Improve mapping core offset to module offset
Core files may not contain `PT_LOAD` segments for read-only VMAs in the
process's address space. When they were elided, our computation of the
offset that any given `PT_LOAD` in the core maps to in the original
module is incorrect. We assume that the earliest `PT_LOAD` segment for
a module in the core file represents the first loadable segment for that
module, but if it doesn't, then our computed offset for a given
`PT_LOAD` segment is off by the size of the elided segments before it.

To fix this, thread through the module's start address from the
`r_debug` information that we extract through `libdw`. In order to
minimize the change surface and avoid making unnecessary changes to
unit tests, make this new parameter optional, and default to computing
it based on an assumption that the first mapping of each module begins
at its load point.

For the tests and implementations that work on core files, pass this new
parameter explicitly, populated from the `r_debug` information.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>