Provide executable in `extract_build_ids` test
Core files can elide read-only segments. If they do, the core file will
be missing the `PHDR` segment, which is needed to locate `PT_DYNAMIC`
and find the `r_debug` information.

`libdw` can handle this, by finding the `PHDR` segment in the executable
image itself instead of in the core file, but it only does this fallback
when the caller explicitly provides the executable to use (it doesn't
use the executable recorded in the core file).

However, pystack's main always passes an executable explicitly before
calling `CoreFileAnalyzer.extract_build_ids`. If the user doesn't
provide the executable, it does an initial pass just to find the
executable, and then a second pass with the executable provided.

Mirror that structure in the test for `extract_build_ids`, ensuring that
an executable is provided before calling it, based on the executable
path recorded in the core file.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>