Avoid confusing anonymous maps for heap maps
The fix for anonymous guard pages was incorrect because `heap_maps`
includes not only maps whose file is listed as `[heap]` but also
anonymous maps following the true heap mapping.

When we're trying to discover which VMAs correspond to the heap, only
consider the subset of `heap_maps` where the path is actually `[heap]`.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>