tests: Drop heap method from smoke tests for 3.11+
In Python 3.11 and later, finding the `PyInterpreterState` by scanning
the heap is unlikely to work for a single-threaded program. In older
Python versions, we should always find a `PyThreadState` on the heap,
which would hold a pointer to the `PyInterpreterState`. From 3.11
onwards, however, the main thread's `PyThreadState` is statically
allocated rather than heap allocated, so this method isn't likely to
work for single threaded programs (though it continues to work for
multithreaded programs, which will have some heap-allocated
`PyThreadState` instances).

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>