Hack around infinite loop while unwinding on musl
The elfutils library fails to successfully unwind multithreaded
applications using musl libc. It gets stuck at the thread entry point
(presumably due to poor debugging information in musl itself) and
continually unwinds to the same instruction pointer forever.

Work around this with an ugly hack: when our callback is passed the same
instruction pointer 5 times in a row, and the instruction pointer before
those 5 was different, check to see if the function we've unwound to
lives in ld-musl.so, and if so, break out of the loop on our own.

This should have minimal performance overhead for typical programs, and
will not affect correctness when unwinding programs using glibc. There
is a risk of false positives if there is some legitimately recursive
function in musl libc.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>