Allow overriding library paths from core files
This enables analyzing a core file even if the libraries that were used
by the application are not installed at the time PyStack is run (perhaps
because the libraries on the machine have been updated). By providing
PyStack with a set of directories to search for libraries in and
placing the versions of the libraries that the application was using in
those directories, users can analyze the core file without needing to
modify the currently installed libraries on the machine.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>