Hack around infinite loop while unwinding on musl
The elfutils library fails to successfully unwind multithreaded
applications using musl libc. It gets stuck at the thread entry point
(presumably due to poor debugging information in musl itself) and
continually unwinds to the same instruction pointer forever.

Work around this with an ugly hack: check the (architecture specific)
stack pointer of each frame. If we ever see both the same instruction
pointer and the same stack pointer for two frames in a row, it means
that the second frame is bogus (since the caller needs to push the
instruction pointer to return to onto the stack when calling
a non-inlined function). If we see a bogus frame, we break out of the
loop on our own.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>