Resolve existing libraries using lib search path
Previously we only used the `--lib-search-path` for libraries that
didn't exist on disk, but we should use it for all libraries.
A different version of a library than the one at its normal location may
exist in `--lib-search-path`, and if it does we should use it rather
than the library at the original file path.

Additionally, only print the "Searching for module" message when we're
actually performing a search.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>