tests: Add hack to PyInstaller test for Alpine
Without this, we fail to find the debug info for the C library that was
copied into the relocated bundle.

That's correct and expected behavior, but it's not useful for testing
that we're correctly remapping libraries, so we'll trick libdw into
loading the split debug from the C library's original location.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>