Drop hard dependency on elfutils 0.188
This version of elfutils isn't available in most distributions yet,
which makes a hard dependency on it (for `dwfl_frame_reg`) annoying.
Since we only need such a new version for trying to avoid the infinite
loop when gathering stacks for interpreters using musl libc, cheat and
avoid performing the check when PyStack itself is built using glibc.

This isn't perfectly correct, but it should work for the overwhelming
majority of cases, and lets us continue to be built with the versions of
elfutils in most common distributions.

Update the README to call out that Alpine users will likely need to
build elfutils from source in order to build PyStack from source, at
least for now.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>