Remove `extras-requires`
Having these dependencies both in `requirements-*.txt` as well as in
`setup.py` allows for inconsistencies, such as the missing `pyinstaller`
dependency that really should be in the test dependencies. After
discussing this, it seems like it's harder to read the `setup.py` file
by Docker and CI in general than the `requirements-*.txt`.

Signed-off-by: Gus Monod <gmonod1@bloomberg.net>