tests: Don't check coverage for dummy programs
The test suite includes several dummy programs whose whole purpose is to
be attached to in order to inspect their state. It's hard to count
coverage for these, because they're run in subprocesses, and because
the interpreter used to run that subprocess might not even be the
interpreter for which `coverage` is installed.

Just exclude those files from coverage reports for now, rather than
trying to find some way to force Coverage to trace them.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>