tests: Correct detection of stripped interpreters
We can't assume that an interpreter is stripped if the string "stripped"
appears in the output of `file`, because (depending on the version)
`file` can report that a particular executable is "not stripped".

Assume that it has symbols if either "stripped" doesn't appear in the
output, or if "not stripped" does.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>