tests: Assume our `/proc/*/maps` regex is correct
Previously we were guarding against the possibility that some lines in
`/proc/$pid/maps` wouldn't match our regex, but that conditional isn't
exercised, and shouldn't be possible unless our regex is incorrect.

If it is incorrect, it would be better to fail the test and find out
about the bad regex than to continue running the test despite the
failure.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>