Add Makefile targets for docker containers (#87)
To allow contributors to more easily use docker to develop and test pystack, add some makefile targets that facilitate building, stoping and running docker containers with an environment where pystack can run.

Additionally, change the Dockerfile to build elfutils from the latest version, as the version shipped in the distribution has a bug that causes some tests to fail. To ensure that the images are as small as possible and that the build is as fast as possible, use multi-stage builds to prepare the elfutils build at the same time the runtime container is being built.