Fix tests in arch Linux
In the latest iteration of the arch Linux container, we need to also
deactivate seccomp to be able to run `gcore` and `ptrace` in the
container.

In addition to that, seems that the new libc is too optimized and some
of our tests are too strict over the number of native frames that must
be present in pthread-based threads, so we need to be a bit more
permissive.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>