ci: Skip aarch64 builds except for new releases
These builds are extremely slow due to needing to rely on emulation, and
we can't adequately test them due to the emulation not providing all
syscalls that we require. There's little point to wasting time building
something that we can't test for every push to a PR branch, so we'll
defer building these until the time a release is cut.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>