Remove `LATEST_VERSION` variable
We failed to keep this up to date when we added support for 3.11, so
clearly this isn't easy to maintain. Instead of having a variable
holding the latest version, make the default branch of the switch fall
through to the latest version. It should be harder to miss updating
this.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>