tests: Fall back to setuptools to build extensions
Prior to Python 3.12, `distutils` was a Python standard library module
that we could use for building extension modules. From 3.12 onwards it
has been removed from the standard library. We can build extension
modules with `setuptools` instead, but we can't easily guarantee that it
will be installed: we run tests against different Python interpreters
than the test suite is being executed by, and can't install dependencies
for those other interpreters.

For now, work around this by installing the `setuptools` package for the
interpreter that runs pytest, and changing the test extension modules to
fall back to `setuptools` if they can't import `distutils`.

This should work as long as any Python 3.12 interpreter that we run
tests against has the `setuptools` package installed.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>