tests: Loosen check for number of eval frames
When running this code under Python 3.12, there are two entry frames
rather than one, and so the test fails. This seems to be a true
difference between 3.11 and 3.12, not us failing to correctly detect the
eval frames, so instead we should loosen the test's assertion. Rather
than checking that the number of native eval frames is exactly 1 for
3.11 and 3.12, check that it's exactly equal to the number of Python
frames marked as entry frames.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>