Handle `PyLongObject` changes in 3.12
The memory layout of a `PyLongObject` has remained the same, but the
field that used to be `Py_ssize_t ob_size` is now `uintptr_t lv_tag`,
and is interpreted differently. Since the offset, size, and alignment of
the old and new field are the same, we can handle this with just
a reinterpret cast of the structure field.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>