Add a LRU cache for live processes
A cache is added to reduce the number of system calls that copy the remote process's memory. Pystack will try to find information about some portion of process memory in the cache, and if there is no information then system call for copying the process memory is called. The LRU cache has a specified capacity.

Signed-off-by: Ivona Stojanovic <istojanovic2@bloomberg.net>