Support version comparisons in `ProcessManager`
Rather than having the `ProcessManager` directly expose the major and
minor version of the process it's managing, have it just expose
a function that can check whether the version is at least as high as
some required version. This turns out to be all we actually need
anywhere in the existing code, and it makes for significantly easier to
read and less error-prone code.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>