Add pkgconfig support in the setup.py (#120)
The pypi.org/project/pkgconfig library provides a way to determine the right compiler and linker flags to use for building against a particular C library. We're not currently leveraging this, and are instead assuming that the elfutils headers and libraries are in the system's default search paths. This commit updates our setup.py to use pkgconfig to locate the right flags for building and linking against libelf and libdw and falls back to the previous behaviour if the pkgconfig query fails.