ci: Automatically publish docs for new releases
On each new release, automatically build our docs and publish them to
GitHub Pages. The docs are also built by our linting workflow, but we
don't attempt to reuse that build here, because the linting build runs
`towncrier` with a phony version number in order to exercise the news
fragments in the news/ directory, and we don't want that fake version in
the changelog we publish.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>