Fix CI target for Fedora
Seems that the latest fedora image ships Python 3.12 with not enough
debugging symbols to provide native information for core files.
Apparently the python3-debug package only provides the debug version of
the runtime and doesn't also provide split-debug information for the
main executable.

To obtain the symbols we need we need to manually install them by
running `dnf debuginfo-install python3`.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>