Implement backoff loop when trying to copy the pthread id
To get the pthread id we copy a buffer of 200 pointers from the
start of the pthread structure and try to match the pid inside. It is
possible that the process doesn't have that much space available in a
valid memory map so reading this much memory will fail.

To fix this, try to reduce the size of the buffer by half every time we
fail to copy from the process until we succeed or we reach 0.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>
Co-authored-by: Matt Wozniski <godlygeek@gmail.com>