Handle threads dying or starting while we attach
Previously, we were fetching the process's list of threads once, and
iterating over that list attaching to each thread. This is racy, though:
after we get the list of threads and before we've suspended every
thread, it's still possible for new threads to start or old threads to
finish.

Handle this by fetching the list of threads at least twice. When we
encounter an error trying to attach to a thread, don't report it unless
the next call to list all of the threads says that thread still belongs
to the process we're attaching to (if the thread has disappeared from
the list, then the reason we failed to attach to that thread is that it
has died).

Similarly, if any new threads show up in the second list of threads,
we'll try to pause those, repeating this until the set of threads that
we've paused contains every thread that the process currently owns.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>