Fetch maps after pausing the process
Previously we were pausing the process only after we had already fetched
the processes mappings, leading to a potential TOCTOU issue: the fetched
mappings might be stale by the time the process was paused and we were
ready to use them.

Handle this by creating a new `ProcessTracer` class responsible for
pausing the process. This class matches the behavior that was previously
found in `BlockingProcessMemoryManager`, except that it's not also an
instance of `AbstractRemoteMemoryManager`: it handles only tracing the
process on construction, and detaching on destruction.

In blocking mode, we create a `ProcessTracer` before reading the
mappings. Later, we pass it to the `ProcessManager`, which takes over
ownership of it, destroying it when we're done analyzing the process and
want to detach.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>