Update `upload-artifact` and `download-artifact`
Both need to be updated at once, and also the uploaded artifacts now
must be each be named something different and are made immutable. This
means that we cannot keep appending to the same archive called
"artifact", and that we instead must create one artifact for the sdist
and one for each combination of CIBW arguments.

Note that the `upload_pypi` step also needs to be changed to ensure it
downloads all of the artifacts and extracts them as before.

Signed-off-by: Gus Monod <gmonod1@bloomberg.net>