Dockerfile: Streamline building elfutils
Build with support for libdebuginfod but not for the debuginfod server.

Ensure the dependencies needed to build against libelf and libdw are
installed in the second stage image.

Install the built library into /usr/local instead of /usr, so that we
don't need to copy as much from the first stage into the second (and
don't risk overwriting something we need). Update `PKG_CONFIG_PATH` so
it can still be located by our build system.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>