docs: Flesh out the core dump tutorial
The example I originally suggested for the core dump tutorial is a bit
too trivial to make an interesting example. Instead, Pablo helped me
come up with an example of a known way to crash the interpreter where
the root cause is not so obvious, which makes a far more engaging
example for readers to dig into and try to make sense of.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>