Add offsets for dict/dictkeys/dictvalues
Using offsets for dict paves the way for handling changes to the size of
PyObject. It's not strictly necessary to make the same change for
dictkeys and dictvalues, but I've done it anyway because it makes the
implementation simpler (decoupling the functions attempting to use the
object from the different structure of that object between versions).

I left `PyDictEntry` alone for now since the layout is still the same
across all Python 3 versions.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>