Fix compiler errors and some error paths
When compiling pystack with gcc 14, the C++ standard library headers are
not giving us transitevely the 'algorithm' include and we are missing it
explicitly so it currently fails to compile.

Additionally, we are not properly handling some of the exceptions that
are being raised when something fails when copying memory when we are
resolving the Python version.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>