Protect against shim frames at the bottom of the stack
It's possible that pystack analyses a core file or a program at a
specific moment where the bottom of the stack has a shim frame but a
normal frame has not been pushed yet or the last normal frame has been
removed because the evaluation loop is returning.

In this case, we were not resolving the code object, but we are not
skipping the frame either because it's the first frame. This was causing
the code object pointer in the frame to be NULL but we are not checking
for that condition when unwinding.

Signed-off-by: Pablo Galindo <pablogsal@gmail.com>