Encapsulate accessing structure fields
Add a new `Structure` class that encapsulates reading and interpreting
the memory of the various Python data structures we need to work with.
This is both easier to work with, and lays the groundwork for us to be
able to handle reading structures without knowing their maximum size at
compile time.

This also fixes several places where we read more bytes than necessary,
which could have resulted in spurious memory read failures due to
reading past the end of a mapping.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>