Fix decoding Python 3.13 stacks without symbols
The eager `isValid` check that we implemented originally fails when
symbols aren't available, as we're unable to look up `PyCode_Type` to
compare the type of `f_executable` against it.

Instead, let's treat every `f_executable` as though it's a `PyCode_Type`
without checking in advance. If it's not able to be interpreted as
a `PyCode_Type`, we'll expect a `RemoteMemCopyError` from chasing an
invalid pointer at some point, and can handle that the way that we have
been handling `!isValid`.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>