fixup! Fix incorrect file offset calculation in memory mapping
Swap the order of operations for relocating the address. It wasn't wrong
as is, but it was unclear - it was subtracting two addresses that are
logically unrelated to each other first, and it might result in an
intermediate integer overflow (which happens to be defined behavior,
because all of the number types involved are unsigned, but is still
weird and surprising).

Changing the order of operations makes it clearer what steps are
actually being taken here: find the offset of the address within the
shared library in the core file, then add that offset to the elf load
point of the shared library.