Handle failures to check if a file is ELF
While attempting to find the executable for a core file, it's possible
that we extract the empty string from the core file where an executable
name ought to have been. In that case we construct a `pathlib.Path` with
the empty string, which happens to behave like a path of `.` - which is
a directory, not a file. This leads to failures in surprising places.

Since the core file may have been generated on a remote machine, or at
a distant point in time, we shouldn't assume that any file name found in
necessarily refers to a file that we can read just because it exists on
disk. It may have been replaced by a directory, or it may have its
permissions set up so that it's not readable by the user we're running
as. Rather than trusting that anything that we can `stat` must be a file
we can read, have our `is_elf` function treat any path from which we
fail to read as though it's not an ELF executable. This lets us continue
processing core files even if files they refer to can't be read by us,
letting us try fallbacks or otherwise attempt to make the best of
things.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>