Fallback to /proc/PID/mem when process_vm_readv not in kernel
Kernels may build with the CONFIG_CROSS_MEMORY_ATTACH configuration
option not set, in this case process_vm_readv returns ENOSYS.

The /proc/PID/mem file in the proc(5) filesystem serves the same purpose
but is less efficient as the data must transfer through the kernel.
We use this as a fallback.

Signed-off-by: Daniel Golding <goldingd89@gmail.com>