Fix handling duplicate symbols when libraries dlopen Python
When libraries like `llvmlite` `dlopen` the Python binary, it creates
duplicate mappings of the binary in the process memory. This caused
PyStack to use the wrong address for `_PyRuntime`, leading to
"Invalid address in remote process" errors.

It happens to work if we stop searching as soon as we find the symbol,
rather than continuing to iterate over other loaded modules searching
for the symbol. This counts on the earliest mapping for the interpreter
binary being the initial mapping made by the loader.

Signed-off-by: Pablo Galindo Salgado <pablogsal@gmail.com>
Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>