Fix a debugging log message
We were incorrectly printing that a 3.13.7 process was 3.13.70 because
we would always print the serial number (always 0 for a final release)
whenever the release level was not a null string - and for final
releases, we had it set to the empty string.

Fix this to check for empty strings, instead, and ensure we never set it
to a null string, instead using a recognizable sentinel for invalid
release levels.

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>